Metrics

+++ api/web/src/components/util/EsriServer.vue
@@ -1,7 +1,7 @@
 <template>
 <div class='border py-2 mx-2'>
     <div class='d-flex'>
-        <h1 class='subheader px-3' v-text='server'></h1>
+        <h1 class='subheader px-3 col-9 text-truncate' v-text='server'></h1>
 
         <div class='ms-auto btn-list mx-3'>
             <RefreshIcon v-if='!disabled && !err && !loading' @click='getList' v-tooltip='"Refresh"' class='cursor-pointer'/>
@@ -13,8 +13,6 @@
 
     <template v-if='err'>
         <Alert title='ESRI Connection Error' :err='err.message' :compact='true'/>
-        <div class="col-md-12 mt-3 pb-2 px-3">
-        </div>
     </template>
     <template v-else-if='loading'>
         <TablerLoading desc='Connecting to ESRI Server'/>
@@ -75,7 +73,7 @@
                                 <MapIcon/><span v-text='lyr.name' class='mx-3'/>
                                 <div class='ms-auto btn-list'>
                                     <CheckIcon v-if='layer && layer.id === lyr.id'/>
-                                    <TablerDelete v-if='!disabled' @delete='deleteLayer' displaytype='icon' label='Delete Layer'/>
+                                    <TablerDelete v-if='!readonly && !disabled' @delete='deleteLayer(lyr)' displaytype='icon' label='Delete Layer'/>
                                 </div>
                             </div>
                         </td>
@@ -111,13 +109,15 @@ export default {
             required: false,
             default: false
         },
+        readonly: {
+            type: Boolean,
+            default: false
+        },
         portal: {
             type: String,
-            required: true,
         },
         token: {
-            type: String,
-            required: true
+            type: Object,
         },
         server: {
             type: String,
@@ -139,7 +139,7 @@ export default {
     watch: {
         layer: function() {
             if (!this.layer) return this.$emit('layer', '');
-            this.$emit('layer', this.stdurl());
+            this.$emit('layer', this.stdurl(true));
         },
         listpath: {
             deep: true,
@@ -157,14 +157,15 @@ export default {
             // TODO Support Directories / Layer Parsing
             postfix = postfix.split('/');
 
-            this.layer = {
-                id: parseInt(postfix[2])
-            };
-
             this.listpath = [{
                 name: postfix[0],
                 type: postfix[1]
             }]
+
+            this.layer = {
+                id: parseInt(postfix[2])
+            };
+
         } else {
             await this.getList();
         }
@@ -188,6 +189,7 @@ export default {
                     return pth.name + '/' + pth.type;
                 }).join('/');
 
+                console.error(layer, listpath);
                 if (!layer || !this.layer) {
                     return this.base + '/rest/services/' + listpath;
                 } else if (layer && this.layer) {
@@ -199,12 +201,15 @@ export default {
 
         },
         createLayer: async function() {
+            if (!this.token) throw new Error('Auth Token is required to create a service');
+
             this.loading = true;
             try {
-                const url = window.stdurl('/api/sink/esri/server/layer');
-                url.searchParams.append('token', this.token);
-                url.searchParams.append('portal', this.portal);
-                url.searchParams.append('server', this.stdurl());
+                const url = window.stdurl('/api/esri/server/layer');
+                url.searchParams.append('token', this.token.token);
+                url.searchParams.append('expires', this.token.expires);
+                if (this.portal) url.searchParams.append('portal', this.portal);
+                url.searchParams.append('server', this.stdurl(false));
 
                 await window.std(url, { method: 'POST' });
 
@@ -213,13 +218,16 @@ export default {
                 this.err = err;
             }
         },
-        deleteLayer: async function() {
+        deleteLayer: async function(layer) {
+            if (!this.token) throw new Error('Auth Token is required to create a service');
+
             this.loading = true;
             try {
-                const url = window.stdurl('/api/sink/esri/server/layer');
-                url.searchParams.append('token', this.token);
-                url.searchParams.append('portal', this.portal);
-                url.searchParams.append('server', this.stdurl());
+                const url = window.stdurl('/api/esri/server/layer');
+                if (this.token) url.searchParams.append('token', this.token.token);
+                if (this.token) url.searchParams.append('expires', this.token.expires);
+                if (this.portal) url.searchParams.append('portal', this.portal);
+                url.searchParams.append('server', this.stdurl(false) + '/' + layer.id);
 
                 await window.std(url, { method: 'DELETE' });
 
@@ -227,16 +235,21 @@ export default {
 
                 await this.getList();
             } catch (err) {
-                this.err = err;
+                this.loading = false;
+                throw err;
             }
         },
         getList: async function() {
             this.loading = true;
             try {
-                const url = window.stdurl('/api/sink/esri/server');
-                url.searchParams.append('token', this.token);
+                const url = window.stdurl('/api/esri/server');
+
+                if (this.token) {
+                    url.searchParams.append('token', this.token.token);
+                    url.searchParams.append('expires', this.token.expires);
+                }
+
                 url.searchParams.append('server', this.stdurl(false));
-                url.searchParams.append('portal', this.portal);
 
                 const res = await window.std(url);
